A bunch of test cases, many of them not passing
Extracted a bunch of test cases from the original paper[1]. A lot of
them are not passing as of now, need various degree of improvement.

Some of the failing tests might actually be handled correctly by the
algorithm and fail due to me picking an alternative, valid syllable
boundaries. These need to be reviewed.

[1] http://www.kalbos.lt/zurnalai/15_numeris/11.pdf